[ISSUE #24] problem biz develop in the second edition of sduoj (#26)
* [ISSUE #24] db model refactor; finish test of most client/manage/judge API

* [ISSUE #24] some problemDescription and problemTag API

* [ISSUE #24] some checkpoint API

+++ sduoj-common/sduoj-common-util/src/main/java/cn/edu/sdu/qd/oj/common/util/UserCacheUtils.java
@@ -25,7 +25,7 @@ public class UserCacheUtils {
     @Autowired
     private RedisUtils redisUtils;
 
-    public String getUsername(int userId) {
+    public String getUsername(long userId) {
         Object o = redisUtils.hget(RedisConstants.REDIS_KEY_FOR_USER_ID_TO_USERNAME, String.valueOf(userId));
         // TODO: 设计本地 Guava 缓存
         return o == null ? null : (String) o;
