+++ sduoj-problem/sduoj-problem-interface/src/main/java/cn/edu/sdu/qd/oj/problem/api/ProblemApi.java
@@ -24,5 +24,5 @@ public interface ProblemApi {
     String SERVICE_NAME = "problem-service";
 
     @GetMapping("/queryIdToTitleMap")
-    Map<Integer, String> queryIdToTitleMap() throws InternalApiException;
+    Map<Long, String> queryIdToTitleMap() throws InternalApiException;
 }
