+++ sduoj-submit/sduoj-submit-interface/src/main/java/cn/edu/sdu/qd/oj/submit/dto/SubmissionListDTO.java
@@ -32,7 +32,7 @@ public class SubmissionListDTO extends BaseDTO {
     @JsonDeserialize(using = HexStringToLongDeserializer.class)
     private Long submissionId;
 
-    private Integer problemId;
+    private Long problemId;
 
     private Integer userId;
