+++ sduoj-submit/sduoj-submit-service/src/main/java/cn/edu/sdu/qd/oj/submit/entity/SubmissionJudgeDO.java
@@ -25,7 +25,7 @@ public class SubmissionJudgeDO extends BaseDO {
     private Long submissionId;
 
     @TableField(SubmissionDOField.PROBLEM_ID)
-    private Integer problemId;
+    private Long problemId;
 
     @TableField(SubmissionDOField.USER_ID)
     private Integer userId;
